fix: convert pino transport to plain JS for worker thread compatibility
